Website controller: wait for deployment to be rolled out
Previously, the controller only waited for the latest generation to be observed and the minimum
number of replicas to be available.
Now it also properly waits until the deployment has been fully rolled out.
